@@ -1,6 +1,6 @@
 [metadata]
 name = edge-tts
-version = 0.0.4
+version = 0.0.5
 author = rany
 author_email = ranygh@riseup.net
 description = Microsoft Edge's TTS
@@ -19,7 +19,7 @@ package_dir=
     = src
 packages = find:
 scripts = bin/edge-playback
-python_requires = >=3.6
+python_requires = >=3.7
 install_requires =
     requests
     websockets>=9.1
